Feature/dialogs (#11)
* Implement confirm dialog for deleting VaultItems.

* Implement delete confirm prompt for groups.

* Implement dialogs to handle all error cases in VaultPageView.

* Add ErrorReportingService to centralize error handling.

* Add error dialogs to all remaining sections of the application.

* Split OverlayPanel modal logic into new Modal templated control.

+++ VaultKeeper.Services/Groups/GroupService.cs
@@ -1,17 +1,24 @@
 ﻿using Microsoft.Extensions.Logging;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using VaultKeeper.Common.Extensions;
 using VaultKeeper.Common.Models.Queries;
 using VaultKeeper.Common.Results;
 using VaultKeeper.Models.Groups;
 using VaultKeeper.Models.Groups.Extensions;
+using VaultKeeper.Models.VaultItems;
 using VaultKeeper.Repositories.Abstractions;
-using VaultKeeper.Services.Abstractions;
+using VaultKeeper.Services.Abstractions.Groups;
 
-namespace VaultKeeper.Services;
+namespace VaultKeeper.Services.Groups;
 
-public class GroupService(IRepository<Group> repository, ILogger<GroupService> logger) : IGroupService
+public class GroupService(
+    IRepository<Group> repository,
+    IRepository<VaultItem> vaultItemRepository,
+    IGroupValidatorService validatorService,
+    ILogger<GroupService> logger) : IGroupService
 {
     public async Task<Result<CountedData<Group>>> GetManyCountedAsync(ReadQuery<Group>? query = null)
     {
@@ -36,7 +43,7 @@ public class GroupService(IRepository<Group> repository, ILogger<GroupService> l
         {
             Group model = group.ToGroup() with { Id = Guid.NewGuid() };
 
-            Result validateResult = await ValidateUpsertAsync(model);
+            Result validateResult = await validatorService.ValidateUpsertAsync(model);
             if (!validateResult.IsSuccessful)
                 return validateResult.WithValue(model).Logged(logger);
 
@@ -64,7 +71,7 @@ public class GroupService(IRepository<Group> repository, ILogger<GroupService> l
             if (existingModel == null)
                 return Result.Failed<Group>(ResultFailureType.NotFound, $"Vault Item ID does not exist: ({group.Id})").Logged(logger);
 
-            Result validateResult = await ValidateUpsertAsync(group);
+            Result validateResult = await validatorService.ValidateUpsertAsync(group);
             if (!validateResult.IsSuccessful)
                 return validateResult.WithValue(group).Logged(logger);
 
@@ -80,15 +87,29 @@ public class GroupService(IRepository<Group> repository, ILogger<GroupService> l
         }
     }
 
-    public async Task<Result> DeleteAsync(Group group)
+    public async Task<Result> DeleteAsync(Group group, CascadeDeleteMode cascadeDeleteMode = CascadeDeleteMode.DeleteChildren)
     {
         logger.LogInformation(nameof(DeleteAsync));
 
         try
         {
-            bool didRemove = await repository.RemoveAsync(group);
-            if (!didRemove)
-                return Result.Failed<Group>(ResultFailureType.NotFound, $"Group ID does not exist: ({group.Id})").Logged(logger);
+            Result validateResult = await validatorService.ValidateDeleteAsync(group, cascadeDeleteMode);
+            if (!validateResult.IsSuccessful)
+                return validateResult.Logged(logger);
+
+            _ = await repository.RemoveAsync(group);
+
+            IEnumerable<VaultItem> childItems = await vaultItemRepository.GetManyAsync(new() { Where = x => x.GroupId == group.Id });
+
+            if (cascadeDeleteMode == CascadeDeleteMode.OrphanChildren)
+            {
+                IEnumerable<KeyValuePair<VaultItem, VaultItem>> updateRequests = childItems.Select(x => new KeyValuePair<VaultItem, VaultItem>(x, x with { GroupId = null }));
+                await vaultItemRepository.UpdateManyAsync(updateRequests);
+            }
+            else
+            {
+                await vaultItemRepository.RemoveManyAsync(childItems);
+            }
 
             return Result.Ok($"Group deleted successfuly (ID: {group.Id}).").Logged(logger);
         }
@@ -96,23 +117,5 @@ public class GroupService(IRepository<Group> repository, ILogger<GroupService> l
         {
             return ex.ToFailedResult().Logged(logger);
         }
-    }
-
-    private async Task<Result> ValidateUpsertAsync(Group group)
-    {
-        if (string.IsNullOrWhiteSpace(group.Name))
-            return Result.Failed(ResultFailureType.BadRequest, "Name is required.");
-
-        bool isDuplicate = await repository.HasAnyAsync(new()
-        {
-            Where = x =>
-                group.Name == x.Name &&
-                group.Id != x.Id
-        });
-
-        if (isDuplicate)
-            return Result.Failed<Group>(ResultFailureType.Conflict, $"Another Group named \"{group.Name}\" already exists.");
-
-        return Result.Ok();
     }
 }
